mqtt_transler/requirements.txt
alembic==1.4.2
altgraph==0.17
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
easygui===0.98.0-unreleased
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
paho-mqtt==1.5.0
pefile==2019.4.18
Pillow==7.2.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
PyMySQL==0.10.0
pypiwin32==223
python-dateutil==2.8.1
python-editor==1.0.4
pywin32==228
pywin32-ctypes==0.2.0
qrcode==6.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
Werkzeug==1.0.1
WMI==1.5.1
WTForms==2.3.3
